backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Security & Encryption
cryptography==42.0.0
python-jose[cryptography]==3.3.0

# AI API Clients
openai==1.10.0
anthropic==0.8.1
httpx==0.26.0

# Export
openpyxl==3.1.2
reportlab==4.0.9

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0
